ast_analysis: Remove unnecessary flag.
I don't think this flag is actually necessary. None of the tests
fail when I remove it, and I never gave it a meaningful name. It
was tracking some characteristic that I thought may be necessary
to get the anchoring rules correct. It should be either removed
or proven to be necessary with a test and then given a better name.