capture: Add special case EXTEND action for states with self-edges.
This is necessary to make the special case in capture5.c (/a(b*)(c)/)
work. Previously, there would be a capture that would start and commit
the capture representing zero-'b' matches in "ac", but "abbbbbbb...c"
would not cause that match to get extended properly, because the
graph search marking paths did not properly handle the self-edge.
Now, a special case flag is set, noting that it needs to add an
EXTEND action for the edge to self.

This may introduce an ambiguity for cases like /a(b*)(b)/, but since
we currently only support putting captures on DFAs, that particular
case may not be possible in practice.