Make fsm_capture_dump() always present, even under -DNDEBUG.
This symbol is kept present to keep the linker scripts happy (i.e. this symbol is listened in the .syms file), because I don't want to take on making some symbols optional when linking.

I decided against making this present but a no-op, because it's harmless to keep the code here when it's not used, and because I think it's less confusing for a caller this way. In other words, I wouldn't want a caller to be able to link against something, call it, and then find there's no outcome.