Add internedstateset.
This wraps creation and modification of state_sets with a cache,
both for existing instances and single-addition transitions between
them.

`fsm_determinise` appears to be spending the majority of its time
saying,
"create '\xfa' -> [123], does it exist? it does, free and reuse that"
"create '\xfb' -> [123], does it exist? it does, free and reuse that"
"create '\xfc' -> [123], does it exist? it does, free and reuse that"
"create '\xfd' -> [123], does it exist? it does, free and reuse that"
....
and moving the cache check into the set construction should eliminate
loads of redundant work. (Combining the numerous `\xaa`..`\xff` edges
that all go to the same state will also help a lot, but is outside the
scope of this.)

The internedstateset is not being used anywhere yet.