Add grouped implementation of edge_set ADT.
This uses a bitset for the labels, since it's very common
to have several labels that lead to the same state (in
particular, "everything else goes here" edges) that could
be beneficially combined, but it's comparatively rare to
have the same label lead to lots of distinct states.

This currently leaves the existing implementation behind an
`#if`, so that we can switch back and forth for benchmarking.

The iterator yields individual <label, state> pairs from each group to
maintain compatibility with the existing implementation, but ideally
some operations (especially `fsm_determinise`) should switch to
processing groups of labels with the same destination together, because
it will massively improve performance with large FSMs.

All tests are passing, including the edge_set properly tests.