Various interface changes driven by edge_set operations that now alloc.
These operations need an allocator handle for the edge_set label bit-set
implementation. These changes should have little impact, because they're
bubbling up an error for allocation failure, which is already largely
non-recoverable, and fsm_removestate appears to only be called in
fsm_mergestates, which is unused (and, per discussion, likely to be
removed outside this PR).