tests/pcre/Makefile: call pcregrep with `--no-jit`. pcregrep bug?
The following tests started failing for me, on `main`:

 - build/tests/pcre-pcregrep/res11:FAIL
 - build/tests/pcre-pcregrep/res12:FAIL
 - build/tests/pcre-pcregrep/res44:FAIL

Sometimes they fail in CI, sometimes they don't. It seems to be
related to `pcregrep`'s `-M` flag: removing the flag causes those
tests to pass, and the failure appears to have to do with
multi-line mode.

@sfstewman found that pcregrep's `-M` option behaves differently
with and without `--no-jit`. Confirmed locally, but I have not
investigated further.