An extra condition to appease ubsan.
This is redundant; it's here to avoid depending on `t->a + 0` where `.a` is NULL. ubsan complains about this, which is probably reasonable. I don't really care if it's reasonable or not; it's no trouble to avoid it in the first place even if I think the extra term here is redundant.