fuzz_re_parser: Add regression test.
This is a regex that (in my development environment) exceeds the 10 msec
time limit for calling `fsm_exec` -- not because evaluating the DFA takes
that long, but our internal `fsm_all`-based check that it's actually a
DFA does. theft found a regex with a disproportionately high number of
edges.