A workaround for the behaviour of $ in single-line mode for PCRE.
From the PCRE spec:
> "The dollar character is an assertion that is true only if the current
> matching point is at the end of the subject string, or immediately before
> a newline at the end of the string (by default). Note, however, that it
> does not actually match the newline."

Here I'm emulating this behaviour by constructing an AST subtree equivalent to `\n?$`.

I hope this is the right way to handle this. Even if it is, this will cause more work in the future, by finding a way to _not_ capture the newline when `$` appears in a capturing group.