Move all of start.c to Rust
The strategy for this commit:

* Implement clear_start / set_start / get_start in Rust as methods,
with argument checks as assertions.

* Implement a C ABI wrapper to match the fsm_clearstart() /
etc. functions.  The non-null argument checks go there.

* Test both sets of functions right there.  We cannot construct a Fsm
in Rust just yet, so the tests hack up just enough initial state for
the code to run; this can be removed later.

* Remove start.c.

Later we can move all of the C API to a c_api.rs or whatever.