fsm_addstate(): no need to initialize has_capture_actions in the realloc block
In all places where a state's .has_capture_actions gets accessed, there is an
assert that the index of the state fits within the fsm->statecount.

Since in the previous commit we initialize the has_capture_actions
field when a new state is added, there is no need to initialize that
field for all the as-yet-unused state structs in the realloc block.