fsm_addstate(): Remove non-working test for statecount == -1
The -1 was meant to be an OOM sentinel, but there is no place in the
code that actually sets fsm->statecount to -1 when the states array
cannot be (re)allocated.

The code after the patch does the realloc(), and properly returns a
failure code if realloc() fails.  I don't think we need to preserve
setting "errno = ENOMEM" since there is no other place in the code
that signals OOM through errno.