Disable multithreading in lx for now.
Running this in a loop:

    $ while true; do sleep 0.1; ./build/bin/lx -l test < src/lx/lexer.lx \
        && echo ok; done

Produces a stream of intermittent failures. There appears to be
unsynchronized access to a shared/global resource, and races on
either `realloc`ing or `free`ing it can nondeterministically lead
to memory corruption.

For now, change to using only one thread -- issue #351 exists as a
reminder to-enable this once the underlying problem has been fixed.