ast_compile: Replace c-stack AST interpreter with one w/ explicit stack.
This lays the groundwork for further changes that could easyl get lost
in the noise.

First, move #includes, enums, struct definitions, and prototypes to an
internal header.

While not yet merged here, evaluating the AST will need to support
forking execution when an AST_EXPR_ALT contains captures that are only
active for some of the subtrees -- a good example of this is the regex
/^(?abc|(ab*c))d$/, which should capture for "ac" or "abbc" but
not for "abc". The metadata necessary to distinguish these (including
when combined with other DFAs) is associated with the end node, so the
entire remainder of execution needs to branch and eventually produce
independent end nodes. It needs to go there so minimisation will
understand that merging them could introduce false matches.

For people familiar with continuations, every ALT subtree with a
distinct set of active captures will need its own continuation, with
that capture set noted on every end node produced during the remainder
of execution.

This commit also splits the start and end nodes into an inner/outer
pair, to explicitly represent what parts of the DFA graph are inside and
outside of match group 0 -- for a regex like /xy/, the input "abcxyz"
should capture as "abc(xy)z". The input's prefix and suffix matched by
the unanchored start and end loops are not included in match group 0.

Otherwise, there should not be any major functional changes here.
There are several pieces related to actual capture management that
will be integrated later on.