Add `fsm_generate_matches` to the public API.
This was implemented specifically for comparison testing against PCRE:
for a given regular expression, build a DFA, generate matching inputs up
to X characters long, and compare their capture behavior between libfsm
and PCRE. It could probably be more generally useful, probably as part
of `fsm`, but that's out of scope right now.

I plan to add generation from a random walk through the DFA later,
and there are a few placeholders for it here.