internedstateset: Reimplement using confluent persistent set.
This adds an alternate implementation of the interned state set
using a confluent persistent set.

This should use *considerably* less memory than the other version,
and has also proven quite a bit faster in benchmarking.

It uses a persistent tree, where subtrees with identical content
are reused, and the tree is rebalanced based on content hashing so
that it converges on the same overall tree structure no matter what
order the individual elements are added (confluence).

See comments for implementation details.