re: Add `-C` flag to disable the `# e.g. "blah"` comments.
`fsm_example`'s current implementation for generating those is very
expensive, so it can be helpful to disable those comments when using re
to test generated inputs (where the example match strings would be
ignored). Disabling `opt.comments` has made a difference of several
seconds for some of my tests.

This does not update the documentation, but some of the other flags also
appear to be missing from the XML, so that should probably be done in
its own pass.