Speed up minimise by working with label sets, not individual labels.
When checking if a label can be used to partition ECs, also note which
other labels led to the same state (and therefore EC). All of those
labels are being checked at once, so they can be skipped in later
iterations. This cuts down on redundant evaluation when large numbers of
labels lead to the same state, which is very common.

This needs an interface change to the edge_set API.