Rename `fuzz` target to `theft`, use `fuzz` for new fuzzer.
There are also some tweaks to the directory-related dependencies in the
Makefiles, because something isn't quite right here, but calling
`theft:: all` was papering over it by way of `all` depending on `mkdir`.
`fuzz` cannot do this because it builds its own object files with
libfuzzer instrumentation, and that includes expecting to provide its
own `main` -- which breaks building `fsm` and the like.

I suspect the real fix involves updating directory dependencies for
other build targets (I was seeing breakage because `src/adt/alloc.c`
could not write into a nonexistent `build/src/adt/` directory), but
that's unrelated to this PR.