Rethinking paths for CI.
I couldn't find a way to cache a repo checked out to `.` so instead I'm checking out the working copy to `wc/`.

Since I'm doing that anyway, I put the pcre suite and build directories adjacent. This way we could have `wc/` and the pcre suite read-only during the build.