determinise: Performance optimization for iterator priority queue.
If we are about to put the current iterator into the priority queue only
to pop it right back out again, note what the next state is on the next
iterator in the queue. Keep directly resume the current iterator as long
as we can, since this saves a lot of time that would otherwise be spent
on pointless queue bookkeeping.

If there's only one iterator remaining, ignore the queue.

Casual benchmarking suggests this can reduce the time spent on
determinisation by about 5-10%.

Note: This moves growing the backing array after the goto label. Since
growing can relocate, refresh the pointer. It also checks if it needs
to grow when switching to the next group.