determinise: Use clearing vector rather than pairwise scan over dst & g.
Previously, we did a scan over `env->groups[]` and `env->dst[]` to pair
up the label groups that should be cleared. This relied on both being
sorted by the destination state (`g->to` and `dst`). We can skip this
pass entirely by saving the node IDs that need to have their label
bitsets cleared later in `env->cvect.ids[]`. In practice this tends to
be much smaller than the entire group set, so there's a small speedup,
and relaxing the ordering constraint also opens up the possibility of
further improvements to the `group_labels_overlap`/`intersect_with`
pass (which dominates the cost of determinisation, and often the overall
runtime).