determinise: Fix typo in parameter array size.
This is a bitset for 0 <= x < 256 stored in a `uint64_t set[4]`,
which was supposed to be represented as 256/64 (to make it explicit
that it was 256 bit split over four 64-bit words), but had a typo
making it actually `uint64_t set[64]` (256/4). This was asymptomatic
since the call to `memcpy` used `sizeof(dst->labels)` anyway, but
gcc 12 rightfully complained about it.