Adds epsilon traversal to the internal fsm_reachable function
fsm_reachable claims to traverse the graph and test if any/all states
match a predicate.  Currently it only traverses non-epsilon edges.

This has the side-effect of causing fsm_empty() to return true if
the start state has only an epsilon edge to the next state.  This is
common in NFAs constructed with re.

Adding epsilon traversal fixes this.