Add minimisation test oracle and minimise_test_case_list.c.
Testing was previously done with `fsm -t equal`, which checks
that the state machine matches the same input as the
expected result, but doesn't confirm that it is minimal. Also
minimised the FSM using a reference implementation and verify
that the resulting state count is as expected.

src/libfsm/minimise_test_oracle.c: Inefficient but easily checked
implementation of minimisation. (This does not handle DFAs with
captures properly.)

tests/minimise/minimise_test_case_list.c: For a list of regex
patterns, check that minimising each ends up with the same number
of states as the test oracle implementation's minimised DFA.

Add an `EXPENSIVE_CHECK` after `fsm_minimise` -- if the FSM does
not have captures, minimise it using the test oracle and check
that the state count matches. Depending on the input size, this
can potentially be quite expensive.

Add `fsm_capture_has_captures`, currently an internal interface.