bugfix: Use EC membership, not state id, for label partitioning check.
When minimisation checked the states' edge sets for sets of labels
leading to particular destinations it was collecting labels that led to
the same state, but it should have instead collected labels to any
states _within the same EC_ (which is progressively refined as
information propagates during minimisation). This could lead to cases
where state(s) would incorrectly split out of an EC, leading to a
non-minimal DFA where some states (each representing an EC set) could be
combined safely and produce a smaller DFA.

A small input that exercised this bug is "ab*c", which produced a
correct but non-minimal DFA with 4 states instead of 3. This was added
to `tests/minimise/minimise_test_case_list.c` as a regression.

Fixes #404.

Rename `edge_set_check_edges` to `edge_set_check_edges_with_EC_mapping`
and note that it's specific to minimisation's use case. Pass in the
current EC map from minimisation and use it to determine which states
should be considered equivalent to the label's first destination state.

Also, update a variable name in a normally compiled-out logging statement
and change `checked_labels`'s length to `[256/64]` to be consistent with
its definition elsewhere.