For VM-generated code, only output variable declarations when we actually use them.
My idea for finding this situation is a little cheesy, with a case like:

    re -k str -r pcre -pl vmc '(.*)(\d*)'

the VM opcode list only contains one instruction, to immediately stop. So I'm detecting that. I think this suffices.

It might be more proper to walk the list of instructions looking for a "fetch", because that's what actually writes to a variable.