-G: Use `c_escputc_str` to escape characters. Respects `-X`.
This adds another callback, `fsm_generate_cb_printf_escaped`, which
uses the current `opt.always_hex` setting.

I don't like it needing to take that struct as a `void *` parameter,
but it needs to be in the public interface so fsm and re can use it.
I'd rather not duplicate the callback into each's `main.c` because
that could lead to surprising discrepancies in formatting if either
is changed.