Missing error handling.
In this case there's no AST to free (the previous code was mistakenly passing NULL to ast_free()), and that's because we jumped to the `error:` handler local to this function. But the parsing does its own error handling, it populates `err` for us, and so using our local handler would be incorrect.