Move the `#include <string.h>` for inline memcmp() calls out to the caller.
I was uncomfortable with this because when a user prints multiple FSM to the same file, they'd be lexing the same header multiple times. But it's also a bug for us where we prefix the generated function with a gcc attribute, and the `#include` gets in the way of that.

The catch is the caller needs to do this, which is what I'm adding for retest.