fuzz/target.c: Add multi-regex and single regex cmp against PCRE.
Add a couple more execution modes to the fuzzer test harness.

In particular, add support for comparing match and capture behavior
against PCRE's -- because this depends on linking with libpcre, it's
disabled by defalut. Set `PCRE_CMP` in the Makefile to enable it, or
pass it as a build argument.