bugfix: resize endid buffer for carry_end_metadata properly.
Previously we doubled the buffer if it wasn't large enough, but
the next endid array may be > twice the size of the old buffer,
so we need to keep expanding until it's large enough.

We weren't saving the updated array size, so this could potentially lead
to repeated doubling and eventually allocation failures.

Also, change the assert for fsm_getendids's result -- if we get to
that point, it should always be found, not just not an insufficient
space error.