Remove nullable ALT backpatching, it's now unreachable.
This was set up to handle edge cases like `^(($)|x)+$` where there is an
alt with some nullable cases with anchors, surrounded by + repetition,
which turned up during fuzzing. This is an obscure edge case, it is
proving very difficult to handle correctly, and there is probably little
value in actually doing so.

Now we are flagging it as an unsupported PCRE construct in
ast_analysis.c's analysis_iter_repetition, so the code using
repeated_alt_backpatches is unreachable. Just remove it.