tests/pcre-anchor/in{99,100}.re: '^(?:$^|a)(?:$^|^)*' and '$[^a]*^|$^'
`^(?:$^|a)(?:$^|^)*`: This should be equivalent to `^(?:^$|a)` because
the first group is always either fully anchored (making the second
redundant) or must match 'a' (making it being anchored at the start
impossible), but that information wasn't being passed properly to the
second group.

`$[^a]*` was returning AST_ANALYSIS_ERROR_UNSUPPORTED_PCRE while
inside the repeat node (`[]*`), so it didn't flag the `^` immediately
after. If a repeat node that can match 0 times is unsupported,
unsatisfiable, etc., just set its max count to 0 and continue. With
this change analysis finds the `^` and treats the whole regex overall
like `$^|$^`, which reduces to `^$`.

The fixes for both of these will be integrated in later commits -- I'm
integrating changes from a working branch that had several intermediate
steps committed to push to another computer for fuzzing.