ast_compile: ALT nodes should pass top-down linking along unmodified.
Lots of other bugs in linking have been indirect consequences of an
ancestor ALT node that changed its top-down x,y linkage to global
or the global start loop instead. Avoiding modification here eliminates
a lot of noise. While trying to simultaneously satisfy the pcre-anchor
tests 92, 99, and 100 I took a step back, thought through the system as
a whole, and realized this was where their linkage was getting messed up.