determinise: Use alt approach for small sort_and_dedup_dst_buf inputs.
I'm still experimenting with the limit, but it's common for
sort_and_dedup_dst_buf to be called with a fairly small input array. The
bitset approach's cost is proportional to the max - min value delta, so
when there's only a few entries it may be faster to do insertion sort
instead -- 16 entries will probably fit in a single cache line.

Changing SMALL_INPUT_LIMIT changes cutoff, commenting it out disables
that approach outright. Setting SMALL_INPUT_CHECK to 1 enables a
pass at the end that verifies the array is actually sorted and unique.

Also, check upfront whether the input array is already sorted and
unique. Experiments so far suggest that this doesn't pay off very often,
but it's very cheap since we're already sweeping over the input to the
min and max values.