re_strings: add support for endids.
This adds an extra parameter to `re_strings_add_str` and
`re_strings_add_raw` that (if non-NULL) will associate a single endid
with the string being added. When `re_strings_build` constructs the
DFA it will produce a separate end state for each end.

This needs further testing with multiple overlapping patterns. When
multiple literal strings appear in the input only the latest match
will be reported.