Add syntax for endids in .fsm format.
These are optional.

I'd expected this to be per state at file scope, but I don't like that the syntax would then allow attempting to attach an id to a non-accepting state. Since we have a block for `end:` already, adding the end ids here means they can't be mixed up.

This looks like:
```
; ./build/bin/re -r pcre -bpl fsm 'ab?c' 'abc*'
0  ->  1 "a"; # e.g. "a"
1  ->  2 "b"; # e.g. "ab"
1  ->  3 "c"; # e.g. "ac"
2  ->  4 "c"; # e.g. "abc"
4  ->  5 "c"; # e.g. "abcc"
5  ->  5 "c"; # e.g. "abcc"

start: 0;
end:   2 = [0], 3 = [0], 4 = [0, 1], 5 = [0];
```