Sort endids written out by fsm_endid_get().
This is for caller convenience, I keep calling qsort for various reasons whenever I call this function, for example to use memcmp on ids.

But the final motivation here was that I wanted to memcmp a const-qualified pointer to an id set stored elsewhere, and I can't do that without allocating for a local copy. And that seems silly when we can just qsort here.