Return bool for match/failure from generated code.
This drops information about which endids match. My intention is to output that later, but I wanted to think about one thing at a time, because this gets confusing. Here I'm making things worse before they get better.

This applies for the codegen that doesn't use the endleaf callbacks. I'm going to call these "default" output.

One observation here is that not all FSMs will have endids attached, and we still need to indicate success/failure. So that's why I'm not using the populated endids to indicate a match.