Use id count to indicate match failure, and >= 0 means success.
The idea here is just to trim down %rt to:
```
%rt = type { ptr, i64 }
```

where we clearly don't need a uint64_t's count of unique ids.

I've purposefully not done the same for single-id interfaces. I don't want to mix success/failure with an id *value*, because the values are opaque (i.e. the meaning of an id value is the responsibility of the caller). Whereas here for AMBIG_MULTIPLE I'm mixing success/failure with the count, not with the id values.

Suggested by @mcy, thank you.